requirements.txt
openai
requests
python-vlc
whisper
pyaudio
python-dotenv
